requirements.txt
itsdangerous==2.0.1
Flask==1.1.2
flask_cors==3.0.10
mongoengine==0.23.1
pymongo[tls,srv]==3.6.1
python_bcrypt==0.3.2